[cmp] Restructured cmp
It was diffcult to use different precision levels
with the way Compare was structured.
This now allows one to define a variable that can
be used for precision specific comparisions through
out an entire package
